+++ tests/Service/MicrosoftGraphBookingServiceTest.php
@@ -10,11 +10,14 @@ use App\Service\MicrosoftGraphHelperService;
 use App\Tests\AbstractBaseApiTestCase;
 use Microsoft\Graph\Http\GraphRequest;
 use Microsoft\Graph\Http\GraphResponse;
+use Psr\Log\LoggerInterface;
 
 class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
 {
     public function testGetBookingData(): void
     {
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
         $microsoftGraphHelperServiceMock = $this->getMockBuilder(MicrosoftGraphHelperService::class)
             ->disableOriginalConstructor()
             ->onlyMethods(['request', 'authenticateAsServiceAccount'])
@@ -35,7 +38,7 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             )
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         $bookingData = $graphService->getBooking('1234');
 
@@ -50,11 +53,13 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
 
     public function testCreateBodyUserId(): void
     {
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
         $microsoftGraphHelperServiceMock = $this->getMockBuilder(MicrosoftGraphHelperService::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         $userId = $graphService->createBodyUserId('useridtest');
 
@@ -63,11 +68,13 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
 
     public function testGetUserBookingFromApiData(): void
     {
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
         $microsoftGraphHelperServiceMock = $this->getMockBuilder(MicrosoftGraphHelperService::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         $data = MicrosoftGraphBookingServiceData::getUserBookingData1();
 
@@ -156,7 +163,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             $resp4,
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         $userBookings = $graphService->getUserBookings('1234567890');
 
@@ -250,7 +259,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             ),
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         // 1. Expired booking.
         $userBooking = $graphService->getUserBookingFromApiData(MicrosoftGraphBookingServiceData::getUserBookingData1());
@@ -364,7 +375,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             ),
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         // 1. Expired
         $userBooking = $graphService->getUserBookingFromApiData(MicrosoftGraphBookingServiceData::getUserBookingData1());
@@ -474,7 +487,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             ),
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         // 1. No conflict
         $data = [
@@ -529,7 +544,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             ),
         );
 
-        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $graphService = new MicrosoftGraphBookingService('test@example.com', 'test', $microsoftGraphHelperServiceMock, $loggerMock);
 
         $data = [
             'name' => 'DOKK1-Lokale-Test1',
@@ -576,7 +593,9 @@ class MicrosoftGraphBookingServiceTest extends AbstractBaseApiTestCase
             ),
         );
 
-        $microsoftGraphService = new MicrosoftGraphBookingService('test', 'test', $microsoftGraphServiceHelperMock);
+        $loggerMock = $this->getMockBuilder(LoggerInterface::class)->getMock();
+
+        $microsoftGraphService = new MicrosoftGraphBookingService('test', 'test', $microsoftGraphServiceHelperMock, $loggerMock);
 
         // Accept conflict.
         $microsoftGraphService->createBookingForResource(