+++ CHANGELOG.md
@@ -9,6 +9,8 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+## [1.1.0]
+
 ### Added
 
 - Added acceptConflict field to resource.